requirements.txt
Django==1.6.1
Pillow==2.3.0
South==0.8.4
argparse==1.2.1
bootstrap-admin==0.3.0
django-countries==1.5
django-filter==0.7
django-grappelli==2.5.1
-e git://github.com/tomviner/django-redmineauth.git@75b0617fdc3705827ba6b918f08d5b890f8d51f4#egg=django_redmineauth-master
djangorestframework==2.3.10
httplib2==0.8
ipython==1.1.0
oauth2==1.5.211
oauthlib==0.6.0
psycopg2==2.5.2
python-dateutil==2.2
python-instagram==0.8.0
python-twitter==1.1
requests==2.1.0
requests-oauthlib==0.4.0
simplejson==3.3.1
six==1.4.1
wsgiref==0.1.2
